+++ hata/discord/auto_moderation/execution_event/tests/test__put_channel_id_into.py
@@ -5,14 +5,14 @@ from ..fields import put_channel_id_into
 
 def test__put_channel_id_into():
     """
-    Tests whether ``put_channel_id_into`` is working as intended.
+    Tests whether ``put_channel_id_into`` works as intended.
     """
     channel_id = 202211160006
     
-    for input_, defaults, expected_output in (
+    for input_value, defaults, expected_output in (
         (0, False, {}),
         (0, True, {'channel_id': None}),
         (channel_id, False, {'channel_id': str(channel_id)}),
     ):
-        data = put_channel_id_into(input_, {}, defaults)
+        data = put_channel_id_into(input_value, {}, defaults)
         vampytest.assert_eq(data, expected_output)