+++ hata/discord/auto_moderation/trigger_metadata/mention_spam.py
@@ -3,7 +3,11 @@ __all__ = ('AutoModerationRuleTriggerMetadataMentionSpam',)
 from scarletio import copy_docs
 
 from .base import AutoModerationRuleTriggerMetadataBase
-from .fields import parse_mention_limit, put_mention_limit_into, validate_mention_limit
+from .constants import AUTO_MODERATION_TRIGGER_MENTION_LIMIT_MAX
+from .fields import (
+    parse_mention_limit, parse_raid_protection, put_mention_limit_into, put_raid_protection_into,
+    validate_mention_limit, validate_raid_protection
+)
 
 
 class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMetadataBase):
@@ -14,29 +18,44 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
     ----------
     mention_limit : `int`
         The amount of mentions in a message after the rule is triggered.
+    raid_protection : `bool`
+        Whether mention raid protection is enabled.
     """
-    __slots__ = ('mention_limit',)
+    __slots__ = ('mention_limit', 'raid_protection')
     
-    def __new__(cls, mention_limit = None):
+    def __new__(cls, mention_limit = ..., *, raid_protection = ...):
         """
         Creates a new mention spam trigger metadata for ``AutoModerationRule``-s.
         
         Parameters
         ----------
-        mention_limit : `None`, `int`, = `None`, Optional
+        mention_limit : `int`, Optional
             The amount of mentions in a message after the rule is triggered.
             
-            Defaults to the allowed maximal amount if given as `None`.
-            
+            Defaults to the allowed maximal amount.
+        raid_protection : `bool`, Optional
+            Whether mention raid protection is enabled.
+        
         Raises
         ------
         TypeError
             - If `mention_limit` type is incorrect.
         """
-        mention_limit = validate_mention_limit(mention_limit)
+        # mention_limit
+        if mention_limit is ...:
+            mention_limit = AUTO_MODERATION_TRIGGER_MENTION_LIMIT_MAX
+        else:
+            mention_limit = validate_mention_limit(mention_limit)
+        
+        # raid_protection
+        if raid_protection is ...:
+            raid_protection = False
+        else:
+            raid_protection = validate_raid_protection(raid_protection)
         
         self = object.__new__(cls)
         self.mention_limit = mention_limit
+        self.raid_protection = raid_protection
         return self
     
     
@@ -48,6 +67,11 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
         repr_parts.append(' mention_limit = ')
         repr_parts.append(repr(self.mention_limit))
         
+        raid_protection = self.raid_protection
+        if raid_protection:
+            repr_parts.append(', raid_protection = ')
+            repr_parts.append(repr(raid_protection))
+        
         repr_parts.append('>')
         return ''.join(repr_parts)
     
@@ -57,6 +81,7 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
     def from_data(cls, data):
         self = object.__new__(cls)
         self.mention_limit = parse_mention_limit(data)
+        self.raid_protection = parse_raid_protection(data)
         return self
     
     
@@ -64,6 +89,7 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
     def to_data(self, *, defaults = False):
         data = {}
         put_mention_limit_into(self.mention_limit, data, defaults)
+        put_raid_protection_into(self.raid_protection, data, defaults)
         return data
     
     
@@ -72,7 +98,12 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
         if type(self) is not type(other):
             return NotImplemented
         
+        # mention_limit
         if self.mention_limit != other.mention_limit:
+            return False
+        
+        # raid_protection
+        if self.raid_protection != other.raid_protection:
             return False
         
         return True
@@ -85,27 +116,30 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
         # mention_limit
         hash_value ^= self.mention_limit
         
+        # raid_protection
+        hash_value ^= self.raid_protection << 6
+        
         return hash_value
     
     
     @copy_docs(AutoModerationRuleTriggerMetadataBase.copy)
     def copy(self):
         new = object.__new__(type(self))
-        
-        # mention_limit
         new.mention_limit = self.mention_limit
-        
+        new.raid_protection = self.raid_protection
         return new
     
     
-    def copy_with(self, *, mention_limit = ..., keyword_presets = ...):
+    def copy_with(self, *, mention_limit = ..., raid_protection = ...):
         """
         Copies the trigger metadata with altering it's attributes based on the given fields.
         
         Parameters
         ----------
-        mention_limit : `None`, `int`, Optional (Keyword only)
+        mention_limit : `int`, Optional (Keyword only)
             The amount of mentions in a message after the rule is triggered.
+        raid_protection : `bool`, Optional
+            Whether mention raid protection is enabled.
         
         Returns
         -------
@@ -122,6 +156,13 @@ class AutoModerationRuleTriggerMetadataMentionSpam(AutoModerationRuleTriggerMeta
         else:
             mention_limit = validate_mention_limit(mention_limit)
         
+        # raid_protection
+        if raid_protection is ...:
+            raid_protection = self.raid_protection
+        else:
+            raid_protection = validate_raid_protection(raid_protection)
+        
         new = object.__new__(type(self))
         new.mention_limit = mention_limit
+        new.raid_protection = raid_protection
         return new