+++ hata/discord/auto_moderation/trigger_metadata/tests/test__AutoModerationRuleTriggerMetadataMentionSpam__constructor.py
@@ -14,6 +14,7 @@ def _assert_is_every_attribute_set(metadata):
     """
     vampytest.assert_instance(metadata, AutoModerationRuleTriggerMetadataMentionSpam)
     vampytest.assert_instance(metadata.mention_limit, int)
+    vampytest.assert_instance(metadata.raid_protection, bool)
 
 
 def test__AutoModerationRuleTriggerMetadataMentionSpam__new__0():
@@ -33,8 +34,13 @@ def test__AutoModerationRuleTriggerMetadataMentionSpam__new__1():
     Case: stuffed.
     """
     mention_limit = 20
+    raid_protection = True
     
-    metadata = AutoModerationRuleTriggerMetadataMentionSpam(mention_limit)
+    metadata = AutoModerationRuleTriggerMetadataMentionSpam(
+        mention_limit = mention_limit,
+        raid_protection = raid_protection,
+    )
     _assert_is_every_attribute_set(metadata)
     
     vampytest.assert_eq(metadata.mention_limit, mention_limit)
+    vampytest.assert_eq(metadata.raid_protection, raid_protection)